requirements.txt
telethon==1.33.1
aiofiles
apscheduler
aiohttp==3.9.3
asyncio==3.4.3
gTTS
dnspython==2.3.0
hachoir
heroku3
motor
psutil
wget
qrcode
future==0.18.3
search_engine_parser
speedtest-cli
py-tgcalls==1.2.9
MukeshAPI
ntgcalls==1.1.2
gitpython
pykeyboard
python-dotenv
pyyaml==6.0.1
tgcrypto
telegraph==2.2.0
yt-dlp
lexica-api==1.4.7
httpx==0.25.2
lyricsgenius
spotipy
SafoneAPI
youtube_search
youtube-search-python==1.6.6
googlesearch-python
unidecode
#git+https://github.com/joetats/youtube_search@master
#git+https://github.com/alexmercerind/youtube-search-python@main
https://github.com/KurimuzonAkuma/pyrogram/archive/v2.2.6.zip
git+https://github.com/AsmSafone/SafoneAPI
bing_image_downloader
openai==0.28
uvicorn
fastapi
pillow
imagehash
# ffmpeg is installed at the system level, but if you want a Python wrapper, uncomment the next line:
# ffmpeg-python
google-generativeai
